backend/services/sreagent/requirements.txt
# Google ADK dependencies
google-adk>=0.1.0
google-genai>=0.1.0

# LiteLLM for OpenAI and other model support
litellm>=1.0.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# MCP client dependencies
mcp>=0.9.0

# Kubernetes client (for direct operations if needed)
kubernetes>=28.0.0

# Database dependencies
sqlalchemy>=2.0.0
pymysql>=1.1.0
alembic>=1.12.0
cryptography>=41.0.0  # For API key encryption

# Authentication dependencies
bcrypt>=4.0.0  # Password hashing
PyJWT>=2.8.0  # JWT token handling
passlib[bcrypt]>=1.7.4  # Password hashing utilities

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
httpx>=0.25.0  # For HTTP MCP transport and cluster inventory client

